+++ src/Cpp/1-getting-started/1-1-3-HelloTriangle-Refactored/HelloTriangleApplication.cpp
@@ -121,7 +121,11 @@ bool HelloTriangleApplication::Load()
         return false;
     }
 
-    _pipeline->SetViewport(0.0f, 0.0f, GetWindowWidth(), GetWindowHeight());
+    _pipeline->SetViewport(
+        0.0f,
+        0.0f,
+        static_cast<float>(GetWindowWidth()),
+        static_cast<float>(GetWindowHeight()));
 
     constexpr VertexPositionColor vertices[] =
     {
