+++ src/Cpp/1-getting-started/1-3-1-ImageLibrary/ImageLibraryApplication.cpp
@@ -146,7 +146,11 @@ bool ImageLibraryApplication::Load()
         return false;
     }
 
-    _pipeline->SetViewport(0.0f, 0.0f, GetWindowWidth(), GetWindowHeight());
+    _pipeline->SetViewport(
+        0.0f,
+        0.0f,
+        static_cast<float>(GetWindowWidth()),
+        static_cast<float>(GetWindowHeight()));
 
     constexpr VertexPositionColorUv vertices[] =
     {
